test/web-platform-tests/to-upstream/svg/types/SVGElement.viewportElement-01.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Element.prototype.hasAttribute</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-element-hasattribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<span data-e2="2" data-F2="3" id="t"></span>

<script>
"use strict";

test(() => {

  const el1 = document.createElementNS("http://www.w3.org/2000/svg", "svg");
  const el2 = document.createElementNS("http://www.w3.org/2000/svg", "svg");

  el1.appendChild(el2);

  assert_equals(el1.viewportElement, null, 'outermost svg element has viewportElement set to null');
  assert_equals(el2.viewportElement, el1);

}, "viewportElement should hold the element which establishes the viewport for the current element");

</script>
